build fix

+++ 3rdparty/openssl-1.0.0d/crypto/ripemd/rmd160.c
@@ -65,7 +65,7 @@
 void do_fp(FILE *f);
 void pt(unsigned char *md);
 #if !defined(_OSD_POSIX) && !defined(__DJGPP__)
-int read(int, void *, unsigned int);
+ssize_t		 read(int, void *, size_t);
 #endif
 
 int main(int argc, char **argv)
